refactor(Seciton): Section 관련 클래스 분리
