+++ docs/Feedback.md
@@ -86,7 +86,7 @@
 - [ ] `LineService`
     - [ ] 원자성과 트랜잭션 학습
         - 메서드가 진행되다가 예기치 못한 상황으로 애플리케이션이 종료된다면?
-    - [ ] `Section`에 대한 책임을 분리
+    - [x] `Section`에 대한 책임을 분리
     - [ ] 변수 이름에 자료형을 사용하면 좋지 않은 이유 (originSectionList)
     - [ ] `deleteSection()`에 대한 테스트
     - [ ] 메서드명 수정
