+++ src/main/java/wooteco/subway/ui/LineController.java
@@ -11,11 +11,9 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import wooteco.subway.dto.LineRequest;
 import wooteco.subway.dto.LineResponse;
-import wooteco.subway.dto.SectionRequest;
 import wooteco.subway.service.LineService;
 
 @RequestMapping("/lines")
@@ -53,18 +51,6 @@ public class LineController {
     @PutMapping("/{id}")
     public ResponseEntity<Void> updateLine(@PathVariable Long id, @RequestBody LineRequest lineRequest) {
         lineService.updateLine(id, lineRequest);
-        return ResponseEntity.ok().build();
-    }
-
-    @PostMapping("/{id}/sections")
-    public ResponseEntity<LineResponse> addSection(@PathVariable Long id, @RequestBody SectionRequest sectionRequest) {
-        lineService.addSection(id, sectionRequest);
-        return ResponseEntity.ok().build();
-    }
-
-    @DeleteMapping("/{id}/sections")
-    public ResponseEntity<Void> deleteLine(@PathVariable Long id, @RequestParam("stationId") Long stationId) {
-        lineService.deleteSection(id, stationId);
         return ResponseEntity.ok().build();
     }
 }
